Privacy Guard: Default WiFi scan OP to "ask" (instead of "allowed")
As the list of available WiFi networks fully qualifies as coarse
location tracking ability, access to it requires therefore the
coarse location permission in android 6 and above.

In android 5 (CM 12.1), the list of available WiFi networks can
be obtained without any permission. The "CM Privacy Guard" has
an AppOp defined, referring to it as "Location", however - it
is simply granted automatically without notification by default!

With this change, Privbacy Guard will prompt the user, if an app
attempts to scan for the available WiFi networks and also labels
this action transparently as "Wi-Fi Scan".

Change-Id: Ibdc45ff5f8da458b0391d8555c939ab1c24f5531