Fix aapt host build issue for libutils
The implementation of the fixes for CVE-2017-0841 and CVE-2017-13323 has
introduced the use of macro 'android_errorWriteLog' into the source code
to build libutils in system/core. This has led to the below error when
building aapt: undefined reference to '__android_log_error_write'

The missing reference (in liblog) is only recognized during build, if
liblog FOLLOWS libutils in the aaptHostStaticLibs string, instead of
preceeding.

Change-Id: I108884d0b1cccefa93c15b9e12d0532ebec3f9e5